requirements.txt
Django == 4.1
psycopg2==2.9.3
django-createsuperuser==2020.12.3
django-reset-migrations==0.4.0
requests==2.27.1
python-dotenv==1.0.0
djangorestframework==3.14
drf-yasg==1.21.6
coreapi==2.3.3
celery==5.2.7
redis==4.3.4